Feature: Github Action

+++ lib/src/feature/ad_panel/presentation/screen/proximity_ad_panels/bloc/proximity_ad_panels_event.dart
@@ -52,7 +52,7 @@ final class UpdateLocationEvent extends ProximityAdPanelsEvent {
 final class UpdateSearchRadiusEvent extends ProximityAdPanelsEvent {
   const UpdateSearchRadiusEvent(this.radiusInKm);
 
-  final double radiusInKm;
+  final int radiusInKm;
 
   @override
   List<Object?> get props => [radiusInKm];
